Use ProtoEqual matcher in EXPECTS
Protobuf does not play well with Go's DeepEqual equality testing that is
used in GoMock.  This PR creates a Protobuf specific equality tester
that uses proto.Equals underneath.
